requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudevents==1.11.0
colorama==0.4.6; platform_system == 'Windows'
db-dtypes==1.3.0; python_version >= '3.7'
deprecation==2.1.0
envelopes==0.4
et-xmlfile==1.1.0; python_version >= '3.6'
flask==3.0.3; python_version >= '3.8'
functions-framework==3.8.1; python_version >= '3.5' and python_version < '4'
google-api-core[grpc]==2.19.2; python_version >= '3.7'
google-auth==2.34.0; python_version >= '3.7'
google-cloud-bigquery==3.25.0; python_version >= '3.7'
google-cloud-core==2.4.1; python_version >= '3.7'
google-cloud-storage==2.18.2; python_version >= '3.7'
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.7.2; python_version >= '3.7'
googleapis-common-protos==1.65.0; python_version >= '3.7'
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
numpy==2.1.0; python_version >= '3.10'
openpyxl==3.1.5; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
proto-plus==1.24.0; python_version >= '3.7'
protobuf==5.28.0; python_version >= '3.8'
pyarrow==17.0.0; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2024.1
requests==2.32.3; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2024.1; python_version >= '2'
urllib3==2.2.2; python_version >= '3.8'
watchdog==5.0.0; python_version >= '3.9'
werkzeug==3.0.4; python_version >= '3.8'
